pyproject.toml
[project]
name = "learning-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-google-genai>=2.1.0",
    "python-dotenv>=1.0.1",
]
